Find JVM and compile the Cython code on Linux.
Now using only Distribute instead of a combination of Distribute and
Setuptools because that seems to be the only way to get Cython
compilation to work. As a side effect, this gets rid of the magic that
added the plugin to nose, so it now has to be added explicitly. We now
provide nosetests.py as a way to do run the tests.