Split up Cython code and revert unit testing.
Split up the Cython code so the the generated C code works on all
operating systems.

Because nosetests.py didn't correctly add the included test.jar to the
class path, go back to the old way of running the unit tests, i.e., add
the plugin to nose using setuptools. (The reason we went away from it
was that we tried to stop using setuptools.)