Call `attach` from `get_env` if necessary.
Have `get_env` call `attach` if it has not been called on the current
thread. This saves the user from having to keep track of which threads
might use the javabridge at some point. And the current
behavior—exploding—is not very useful.

We don't call `detach` for the user. It is possible that not calling
`detach` will prevent the JVM from exiting on some operating
systems, or that it will waste resources on the Java side if you have
thousands of ephemeral threads.