Fix find_javahome for Python 3 on macOS
In Python 3, the path returned by the subprocess call to
`/usr/libexec/java_home` is a bytes, not an str which `os.path.join`
expects. The solution is to decode the path. Special consideration must
be made for Python 2, because `libc.dlopen_preflight` expects bytes/str,
so the path must be re-encoded.