Dynamically load libjvm.so before importing _javabridge
.. as it is very unlikely that the machine you run javabridge on
is the same machine that you built javabridge on, and it is also
unlikely even in that case that you would not update java on the
same machine at some point.

This is done using a function in locate to find the jvm shared
library.  This function is also reused in setup.py so that the
same jvm shared library gets found in the same way at build and
also at run-time (on the same machine).